% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{add_proportion}
\alias{add_proportion}
\title{Adds a population_proportion column to a poststratification frame.}
\usage{
add_proportion(
  poststratification_frame,
  model_variables,
  estimates_by,
  weight_column
)
}
\arguments{
\item{weight_column}{}
}
\description{
Adds a population_proportion column to a poststratification frame.
}
