% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify_frequentist.R
\name{all_strata_results_to_more_results}
\alias{all_strata_results_to_more_results}
\title{all_strata_results_to_more_results}
\usage{
all_strata_results_to_more_results(strata_results, frame, estimates_by)
}
\arguments{
\item{estimates_by}{}
}
\description{
all_strata_results_to_more_results
}
