% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_poststratification.R
\name{calculate_raised_totals}
\alias{calculate_raised_totals}
\title{Calculate raised totals}
\usage{
calculate_raised_totals(
  survey,
  survey_strata_weights,
  strata,
  results_by,
  dependent_variable
)
}
\arguments{
\item{dependent_variable}{}
}
\value{

}
\description{
Calculate raised totals
}
