% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_poststratification.R
\name{calculate_survey_strata_weights}
\alias{calculate_survey_strata_weights}
\title{Calculate survey strata weights}
\usage{
calculate_survey_strata_weights(
  survey,
  population_dataset,
  strata,
  population_weight_column
)
}
\arguments{
\item{population_dataset}{}
}
\description{
Calculate survey strata weights
}
