% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{check_character_variable_values}
\alias{check_character_variable_values}
\title{Check character variable values}
\usage{
check_character_variable_values(model, frame)
}
\arguments{
\item{frame}{}
}
\description{
Check character variable values
}
