% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{check_frame_model_alignment}
\alias{check_frame_model_alignment}
\title{Check frame-model alignment}
\usage{
check_frame_model_alignment(model, frame)
}
\arguments{
\item{frame}{}
}
\description{
Check frame-model alignment
}
