% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{check_grouping_variable_values}
\alias{check_grouping_variable_values}
\title{Check grouping variable values}
\usage{
check_grouping_variable_values(model, frame)
}
\arguments{
\item{frame}{}
}
\description{
Check grouping variable values
}
