% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{collect_draws_by_strata}
\alias{collect_draws_by_strata}
\title{Collect draws by strata}
\usage{
collect_draws_by_strata(draws, strata_prediction, estimates_by)
}
\arguments{
\item{estimates_by}{}
}
\description{
Usually used internally by poststratify. Provides draws, grouped by
}
