% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{create_poststratification_frame}
\alias{create_poststratification_frame}
\title{Create poststratification frame from a population level survey such as a census}
\usage{
create_poststratification_frame(census, strata_variables, weight_column = 1)
}
\arguments{
\item{weight_column}{}
}
\value{

}
\description{
Create poststratification frame from a population level survey such as a census
}
