% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_census}
\alias{example_census}
\title{Example Census}
\format{
A data frame
\describe{
  \item{region}{Regions of a country represented as letters}
  \item{age_group}{Age group}
  \item{ethnicity}{Ethnicity}
  \item{gender}{Gender}
  \item{n}{Number of people in this category in the country}
  ...
}
}
\source{
Simulated in package
}
\usage{
example_census
}
\description{
A simulated dataset for use as an example census
}
\keyword{datasets}
