% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_survey}
\alias{example_survey}
\title{Example Survey}
\format{
A data frame
\describe{
  \item{region}{Regions of a country represented as letters}
  \item{age}{Age in years}
  \item{ethnicity}{Ethnicity}
  \item{non_negative_response}{A non-negative response to a question e.g. the amount of money they make}
  \item{binary_response}{A binary response to a question e.g. if they play golf}
  ...
}
}
\source{
Simulated in package
}
\usage{
example_survey
}
\description{
A simulated dataset for use as an example survey
}
\keyword{datasets}
