% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{get_character_variable_values}
\alias{get_character_variable_values}
\title{Get character variable values}
\usage{
get_character_variable_values(model, frame)
}
\arguments{
\item{frame}{}
}
\description{
Get character variable values
}
