% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_independent_variables}
\alias{get_independent_variables}
\title{Get dependent variables from a brms model}
\usage{
get_independent_variables(model)
}
\arguments{
\item{model}{}
}
\description{
Get dependent variables from a brms model
}
