% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_poststratified_draws}
\alias{get_poststratified_draws}
\title{Create post-stratified draws from a model}
\usage{
get_poststratified_draws(model, new_data, weight_column = population_total)
}
\arguments{
\item{poststratification_frame}{}
}
\value{

}
\description{
Create post-stratified draws from a model
}
