% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_poststratified_estimates}
\alias{get_poststratified_estimates}
\title{Title}
\usage{
get_poststratified_estimates(
  model,
  new_data,
  model_variables,
  estimates_by,
  weight_column = population_total,
  lower_confidence = 0.025,
  upper_confidence = 1 - lower_confidence
)
}
\arguments{
\item{upper_confidence}{}
}
\value{

}
\description{
Title
}
