% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify_frequentist.R
\name{get_results_for_all_strata}
\alias{get_results_for_all_strata}
\title{get_results_for_all_strata}
\usage{
get_results_for_all_strata(
  model,
  poststratification_frame,
  slice_size = NULL,
  ...
)
}
\arguments{
\item{poststratification_frame}{}
}
\description{
get_results_for_all_strata
}
