% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{perform_poststratification}
\alias{perform_poststratification}
\title{Perform poststratification}
\usage{
perform_poststratification(
  model,
  reduced_frame,
  estimates_by,
  lower_confidence,
  upper_confidence,
  draws,
  ...
)
}
\arguments{
\item{estimate_value}{}
}
\description{
Perform poststratification
}
