% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{poststratify}
\alias{poststratify}
\title{Get poststratified estimates from a model}
\usage{
poststratify(
  model,
  poststratification_frame,
  estimates_by,
  weight_column = "n",
  lower_confidence = 0.025,
  upper_confidence = 1 - lower_confidence,
  allow_new_levels = FALSE,
  draws = "predicted"
)
}
\arguments{
\item{draws}{predicted (default) or fitted. predicted incorporates all uncertainty.}
}
\value{

}
\description{
Get poststratified estimates from a model
}
