% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify_frequentist.R
\name{poststratify_frequentist}
\alias{poststratify_frequentist}
\title{Poststratify frequentist}
\usage{
poststratify_frequentist(
  model,
  poststratification_frame,
  estimates_by,
  weight_column = n,
  lower_confidence = 0.025,
  upper_confidence = 1 - lower_confidence,
  draws = "predicted"
)
}
\arguments{
\item{draws}{}
}
\description{
Poststratify frequentist
}
