% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{poststratify}
\alias{poststratify}
\title{Get poststratified estimates from a model}
\usage{
poststratify(
  model,
  poststratification_frame,
  estimates_by,
  weight_column = n,
  lower_confidence = 0.025,
  upper_confidence = 1 - lower_confidence,
  draws = "predicted",
  large_frame = FALSE,
  progress = FALSE,
  ...
)
}
\arguments{
\item{draws}{predicted (default) or fitted. Predicted incorporates all uncertainty.}

\item{...}{Extra arguments for adding draws from tidybayes. See tidybayes::add_predicted_draws() for details.}
}
\description{
Get poststratified estimates from a model
}
